Created azure_messaging_queues crate from azure_storage::queue
