+++ sdk/messaging_queues/src/number_of_messages.rs
