+++ sdk/messaging_queues/src/pop_receipt.rs
