+++ sdk/messaging_queues/src/prelude.rs
@@ -1,4 +1,4 @@
-pub use crate::queue::clients::{AsPopReceiptClient, AsQueueClient};
+pub use crate::clients::{AsPopReceiptClient, AsQueueClient};
 pub use crate::{
     MessageTTL, NumberOfMessages, PopReceipt, QueueClient, QueueStoredAccessPolicy,
     VisibilityTimeout,
