+++ sdk/messaging_queues/src/requests/create_queue_builder.rs
@@ -1,53 +1,46 @@
-use crate::queue::clients::QueueClient;
-use crate::queue::prelude::*;
-use crate::queue::responses::*;
+use crate::clients::QueueClient;
+use crate::responses::*;
 use azure_core::headers::add_optional_header;
 use azure_core::prelude::*;
 use std::convert::TryInto;
 
 #[derive(Debug, Clone)]
-pub struct GetMessagesBuilder<'a> {
+pub struct CreateQueueBuilder<'a> {
     queue_client: &'a QueueClient,
-    number_of_messages: Option<NumberOfMessages>,
-    visibility_timeout: Option<VisibilityTimeout>,
     timeout: Option<Timeout>,
+    metadata: Option<&'a Metadata>,
     client_request_id: Option<ClientRequestId<'a>>,
 }
 
-impl<'a> GetMessagesBuilder<'a> {
+impl<'a> CreateQueueBuilder<'a> {
     pub(crate) fn new(queue_client: &'a QueueClient) -> Self {
-        GetMessagesBuilder {
+        CreateQueueBuilder {
             queue_client,
-            number_of_messages: None,
-            visibility_timeout: None,
             timeout: None,
+            metadata: None,
             client_request_id: None,
         }
     }
 
     setters! {
-        number_of_messages: NumberOfMessages => Some(number_of_messages),
-        visibility_timeout: VisibilityTimeout => Some(visibility_timeout),
+        metadata: &'a Metadata => Some(metadata),
         timeout: Timeout => Some(timeout),
         client_request_id: ClientRequestId<'a> => Some(client_request_id),
     }
 
     pub async fn execute(
         &self,
-    ) -> Result<GetMessagesResponse, Box<dyn std::error::Error + Sync + Send>> {
-        let mut url = self.queue_client.url_with_segments(Some("messages"))?;
+    ) -> Result<CreateQueueResponse, Box<dyn std::error::Error + Sync + Send>> {
+        let mut url = self.queue_client.url_with_segments(None)?;
 
-        self.visibility_timeout.append_to_url_query(&mut url);
-        self.number_of_messages.append_to_url_query(&mut url);
         self.timeout.append_to_url_query(&mut url);
-
-        debug!("url == {}", url);
 
         let request = self.queue_client.storage_client().prepare_request(
             url.as_str(),
-            &http::method::Method::GET,
+            &http::method::Method::PUT,
             &|mut request| {
                 request = add_optional_header(&self.client_request_id, request);
+                request = add_optional_header(&self.metadata, request);
                 request
             },
             None,
@@ -58,7 +51,7 @@ impl<'a> GetMessagesBuilder<'a> {
             .storage_client()
             .storage_account_client()
             .http_client()
-            .execute_request_check_status(request.0, http::status::StatusCode::OK)
+            .execute_request_check_status(request.0, http::status::StatusCode::CREATED)
             .await?;
 
         Ok((&response).try_into()?)
