+++ sdk/messaging_queues/src/requests/get_queue_acl_builder.rs
@@ -1,19 +1,21 @@
-use crate::queue::clients::QueueClient;
-use crate::queue::responses::*;
+use crate::clients::QueueClient;
+use crate::responses::*;
 use azure_core::headers::add_optional_header;
 use azure_core::prelude::*;
+use http::method::Method;
+use http::status::StatusCode;
 use std::convert::TryInto;
 
 #[derive(Debug, Clone)]
-pub struct DeleteQueueBuilder<'a> {
+pub struct GetQueueACLBuilder<'a> {
     queue_client: &'a QueueClient,
     timeout: Option<Timeout>,
     client_request_id: Option<ClientRequestId<'a>>,
 }
 
-impl<'a> DeleteQueueBuilder<'a> {
+impl<'a> GetQueueACLBuilder<'a> {
     pub(crate) fn new(queue_client: &'a QueueClient) -> Self {
-        DeleteQueueBuilder {
+        Self {
             queue_client,
             timeout: None,
             client_request_id: None,
@@ -27,14 +29,18 @@ impl<'a> DeleteQueueBuilder<'a> {
 
     pub async fn execute(
         &self,
-    ) -> Result<DeleteQueueResponse, Box<dyn std::error::Error + Sync + Send>> {
+    ) -> Result<GetQueueACLResponse, Box<dyn std::error::Error + Sync + Send>> {
         let mut url = self.queue_client.url_with_segments(None)?;
+
+        url.query_pairs_mut().append_pair("comp", "acl");
 
         self.timeout.append_to_url_query(&mut url);
 
+        trace!("url == {}", url);
+
         let request = self.queue_client.storage_client().prepare_request(
             url.as_str(),
-            &http::method::Method::DELETE,
+            &Method::GET,
             &|mut request| {
                 request = add_optional_header(&self.client_request_id, request);
                 request
@@ -47,7 +53,7 @@ impl<'a> DeleteQueueBuilder<'a> {
             .storage_client()
             .storage_account_client()
             .http_client()
-            .execute_request_check_status(request.0, http::status::StatusCode::NO_CONTENT)
+            .execute_request_check_status(request.0, StatusCode::OK)
             .await?;
 
         Ok((&response).try_into()?)
