+++ sdk/messaging_queues/src/requests/put_message_builder.rs
@@ -1,5 +1,5 @@
-use crate::queue::responses::*;
-use crate::queue::*;
+use crate::prelude::*;
+use crate::responses::*;
 use azure_core::headers::add_optional_header;
 use azure_core::prelude::*;
 use std::convert::TryInto;
