+++ sdk/messaging_queues/src/responses/delete_message_response.rs
