+++ sdk/messaging_queues/src/responses/delete_queue_response.rs
