+++ sdk/messaging_queues/src/responses/get_queue_acl_response.rs
@@ -1,6 +1,7 @@
-use crate::{QueueStoredAccessPolicy, StoredAccessPolicyList};
+use crate::QueueStoredAccessPolicy;
 use azure_core::headers::CommonStorageResponseHeaders;
 use azure_core::PermissionError;
+use azure_storage::StoredAccessPolicyList;
 use bytes::Bytes;
 use http::response::Response;
 use std::convert::TryInto;
