+++ sdk/messaging_queues/src/responses/get_queue_metadata_response.rs
