+++ sdk/messaging_queues/src/responses/put_message_response.rs
@@ -1,5 +1,5 @@
-use crate::xml::read_xml;
 use azure_core::headers::{utc_date_from_rfc2822, CommonStorageResponseHeaders};
+use azure_storage::xml::read_xml;
 use bytes::Bytes;
 use chrono::{DateTime, Utc};
 use http::response::Response;
