+++ sdk/messaging_queues/src/responses/set_queue_metadata_response.rs
