+++ sdk/storage/src/queue/mod.rs
@@ -1,18 +1,4 @@
-mod clients;
-mod message_ttl;
-mod number_of_messages;
-mod pop_receipt;
-pub mod prelude;
 mod queue_service_properties;
-mod queue_stored_access_policy;
 pub mod requests;
 pub mod responses;
-mod visibility_timeout;
-
-pub use clients::*;
-pub use message_ttl::MessageTTL;
-pub use number_of_messages::NumberOfMessages;
-pub use pop_receipt::PopReceipt;
 pub use queue_service_properties::QueueServiceProperties;
-pub use queue_stored_access_policy::QueueStoredAccessPolicy;
-pub use visibility_timeout::VisibilityTimeout;
