+++ sdk/storage/src/queue/requests/set_queue_service_properties_builder.rs
@@ -1,6 +1,6 @@
 use crate::core::clients::StorageClient;
 use crate::queue::responses::*;
-use crate::QueueServiceProperties;
+use crate::queue::QueueServiceProperties;
 use azure_core::headers::add_optional_header;
 use azure_core::prelude::*;
 use std::convert::TryInto;
