+++ sdk/storage/src/queue/responses/mod.rs
@@ -1,32 +1,8 @@
-mod clear_messages_response;
-mod create_queue_response;
-mod delete_message_response;
-mod delete_queue_response;
-mod get_messages_response;
-mod get_queue_acl_response;
-mod get_queue_metadata_response;
 mod get_queue_service_properties_response;
 mod get_queue_service_stats_response;
 mod list_queues_response;
-mod peek_messages_response;
-mod put_message_response;
-mod set_queue_acl_response;
-mod set_queue_metadata_response;
 mod set_queue_service_properties_response;
-mod update_message_response;
-pub use clear_messages_response::ClearMessagesResponse;
-pub use create_queue_response::CreateQueueResponse;
-pub use delete_message_response::DeleteMessageResponse;
-pub use delete_queue_response::DeleteQueueResponse;
-pub use get_messages_response::GetMessagesResponse;
-pub use get_queue_acl_response::GetQueueACLResponse;
-pub use get_queue_metadata_response::GetQueueMetadataResponse;
 pub use get_queue_service_properties_response::GetQueueServicePropertiesResponse;
 pub use get_queue_service_stats_response::GetQueueServiceStatsResponse;
 pub use list_queues_response::ListQueuesResponse;
-pub use peek_messages_response::PeekMessagesResponse;
-pub use put_message_response::PutMessageResponse;
-pub use set_queue_acl_response::SetQueueACLResponse;
-pub use set_queue_metadata_response::SetQueueMetadataResponse;
 pub use set_queue_service_properties_response::SetQueueServicePropertiesResponse;
-pub use update_message_response::UpdateMessageResponse;
